Split `LightResize.cs` into one file per type
